added new math module to preprocess packages and removed numba from requirements.txt
